Reorganize lots of code because my eyes bleed sometimes :wrench:
