+++ web-app/src/com/lewismorgan/web/ContactComponent.kt
@@ -6,7 +6,6 @@ import react.RBuilder
 import react.RComponent
 import react.RProps
 import react.RState
-import react.dom.div
 
 class ContactComponent : RComponent<RProps, RState>() {
   override fun RBuilder.render() {
