+++ web-app/src/com/lewismorgan/web/EducationComponent.kt
@@ -5,7 +5,6 @@ import react.RBuilder
 import react.RComponent
 import react.RProps
 import react.RState
-import react.dom.div
 
 class EducationComponent : RComponent<RProps, RState>() {
   override fun RBuilder.render() {
