+++ web-app/src/com/lewismorgan/web/bootstrap-custom.css
@@ -1,9 +1,12 @@
+/*
+  Contains all Bootstrap CSS Classes that changes the layout/styles
+ */
+
 @media (min-width: 992px) {
   header.masthead {
     padding-top: calc(6rem + 106px);
     padding-bottom: 6rem;
   }
-
 }
 
 header.masthead {
@@ -38,35 +41,11 @@ header.masthead {
 
 .carousel-icon-item {
   margin-top: 8rem;
-  /* position: absolute; */
-  /* right: inherit; */
-  /* left: inherit; */
   min-width: 100%;
-  /* height: 20rem; */
   text-align: center;
   font-size: 12em;
 }
 
-.list-stack {
-  list-style: none;
-  padding-left: 0;
-  font-size: 1.25rem
-}
-
-.floating-item-container {
-  position: fixed;
-  z-index: 9999;
-}
-
-.jump-top-nav-btn {
-  position: inherit;
-  bottom: 5rem;
-  right: 1rem;
-  height: 5rem;
-  width: 5rem;
-  font-size: large;
-}
-
-.list-stack-item {
-
+.list-inline-item a {
+  color: inherit
 }
