+++ web-app/src/com/lewismorgan/web/bootstrap/nav/NavComponent.kt
@@ -58,6 +58,7 @@ class NavComponent : RComponent<NavProps, NavState>() {
   }
 
   private fun onSelectNavItem(index: Int): (Event) -> Unit = {
+    console.log("NAV ITEM SELECTED")
     setState {
       activeIndex = index
     }
