+++ web-app/src/com/lewismorgan/web/bootstrap/nav/NavItemComponent.kt
@@ -1,11 +1,11 @@
 package com.lewismorgan.web.bootstrap.nav
 
-import com.lewismorgan.web.misc.getSmoothScrollingHandler
-import com.lewismorgan.web.misc.navHashLink
+import kotlinx.html.js.onClickFunction
 import kotlinx.html.role
 import org.w3c.dom.events.Event
 import react.RBuilder
 import react.RComponent
+import react.RHandler
 import react.RProps
 import react.RState
 import react.ReactElement
@@ -13,8 +13,6 @@ import react.dom.li
 
 interface NavItemProps : RProps {
   var isActive: Boolean
-  var href: String
-  var text: String
   var onSelect: (Event) -> Unit
 }
 
@@ -25,19 +23,10 @@ class NavItemComponent : RComponent<NavItemProps, RState>() {
   override fun RBuilder.render() {
     li("nav-item ${if (props.isActive) "active" else ""}") {
       attrs.role = "presentation"
-      // TODO: NavHashLinks broken for active status see https://github.com/rafrex/react-router-hash-link/issues/29
-      navHashLink("/" + props.href, onClick = props.onSelect, className = "nav-link", activeClassName = "active", scroll = getSmoothScrollingHandler()) {
-        +props.text
-      }
+      attrs.onClickFunction = props.onSelect
+      children()
     }
   }
 }
 
-fun RBuilder.navigationItem(href: String, text: String, onSelect: ((Event) -> Unit)? = null): ReactElement {
-  return child<NavItemProps, NavItemComponent> {
-    if (onSelect != null)
-      attrs.onSelect = onSelect
-    attrs.href = href
-    attrs.text = text
-  }
-}
+fun RBuilder.navigationItem(handler: RHandler<NavItemProps>): ReactElement = child<NavItemProps, NavItemComponent>(handler)
