+++ web-app/src/com/lewismorgan/web/components.css
@@ -1,0 +1,25 @@
+/* Contains all CSS specific for the site */
+
+.floating-item-container {
+  position: fixed;
+  z-index: 9999;
+}
+
+.jump-top-nav-btn {
+  position: inherit;
+  bottom: 5rem;
+  right: 1rem;
+  height: 5rem;
+  width: 5rem;
+  font-size: large;
+}
+
+.list-stack {
+  list-style: none;
+  padding-left: 0;
+  font-size: 1.25rem
+}
+
+.list-stack-item {
+
+}
