+++ web-app/src/index/index.css
@@ -6,10 +6,6 @@
   color: inherit;
 }
 
-.list-inline-item a {
-  color: inherit
-}
-
 .facebook-icon:hover {
   color: #3B5998
 }
